legal_rag_system/requirements.txt
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0
qdrant-client>=1.6.0
langchain>=0.0.340
langchain-community>=0.0.10
pypdf>=3.17.0
pymupdf>=1.23.0
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.3.0
requests>=2.31.0
aiohttp>=3.9.0
redis>=5.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
pandas>=2.0.0
tiktoken>=0.5.0
sse-starlette>=1.6.5